refactor(claw-server-rust): split typed route families
